Read manifest from variables
